Better format=flowed text formatter
This works from the DOM, which is a bit of a problem.  I can't figure
out how to create a DOM.Document with WebKit (it doesn't have a
DOMParser), so we have to pass the DOM until we convert to text.  With
my "HTML everywhere" crusade, this means the DOM goes all the way out to
the RFC822.Message, which is a bit silly.

The new function html_to_flowed_text is destructive to the DOM, which
might be weird.  It's okay for now, since we're only calling it as we're
closing the associated WebView.  It could be fixed to restore the DOM
before exiting.

There are some remnants of other attempts included, in case I want to go
back to them.  And there's a lot of cruft to clean up.