WIP - Destroying the ComposerWidget
The should_close() method was a bit strange -- sometimes it returned
true, indicating that the caller should destroy the widget, sometimes it
returned false, indicating that the widget should remain open, and
sometimes it returned false, indicating that the widget would destoy
itself shortly.  With the widget living inside a container that has to
figure out what it should be doing to itself, this is a bit confusing.

So, we change this to a try_close() method, which will ensure that
close() is called (eventually) if needed.  This in turn calls the
close() method of the container, which should destroy the widget.  But
when the Embed tries to destroy the widget, we get into a loop of
widget.destroy() calls, for some reason.